Improve overwriting files
Previously, if you tried to write a new file, while overwriting a
previous one, it would return the FILE_ALREADY_EXISTS error if the
filesizes weren't the same.

This doesn't make sense, espescially since the user explititly set the
overwrite flag true.

This new behaviour returns FILE_ALREADY_EXISTS only if overwrite is
false.

If overwrite is true and it encounters an identically named file, it
either uses it (if the sizes are identical) or deletes it and moves on.

(cherry picked from commit f62b6cb050f94f5e71fc10d63834acd2827a75f0)